IBM-Swift/Kitura#773 Moved RedisInfo to its own file
