+++ Sources/Redis.swift
@@ -2529,71 +2529,3 @@ public class Redis {
         return createError(redisError, code: 1)
     }
 }
-
-public struct RedisInfo {
-    
-    public let server: RedisInfoServer
-    public let client: RedisInfoClient
-
-    public init(_ redisReply: RedisString) {
-        
-        let convertedStr = redisReply.asString
-        let newline = "\r\n"
-        let strArray = convertedStr.components(separatedBy: newline)
-        var parsedInfo: [String: String] = [:]
-        
-        for val in strArray {
-            let pos = val.range(of: ":")
-            if let pos = pos {
-                parsedInfo[val.substring(to: pos.lowerBound)] = val.substring(from: pos.upperBound)
-            }
-        }
-        
-        server = RedisInfoServer(parsedInfo)
-        client = RedisInfoClient(parsedInfo)
-    }
-    
-    public struct RedisInfoClient {
-        public let connected_clients: Int
-        public let blocked_clients: Int
-
-        fileprivate init(_ redisInfo: [String: String]) {
-            self.connected_clients = Int(redisInfo["connected_clients"]!)!
-            self.blocked_clients = Int(redisInfo["blocked_clients"]!)!
-        }
-    }
-    
-    public struct RedisInfoServer {
-        
-        public let redis_version: String
-        public let redis_mode: String
-        public let os: String
-        public let arch_bits: Int
-        public let process_id: Int
-        public let tcp_port: Int
-        public let uptime_in_seconds: Int
-        public let uptime_in_days: Int
-        
-        fileprivate init(_ redisInfo: [String: String]) {
-            self.redis_version = redisInfo["redis_version"]!
-            self.redis_mode = redisInfo["redis_mode"]!
-            self.os = redisInfo["os"]!
-            self.arch_bits = Int(redisInfo["arch_bits"]!)!
-            self.process_id = Int(redisInfo["process_id"]!)!
-            self.tcp_port = Int(redisInfo["tcp_port"]!)!
-            self.uptime_in_seconds  = Int(redisInfo["uptime_in_seconds"]!)!
-            self.uptime_in_days  = Int(redisInfo["uptime_in_days"]!)!
-        }
-        
-        public func checkVersionCompatible(major: Int, minor: Int=0) -> Bool{
-            let v = self.redis_version.components(separatedBy: ".")
-            return Int(v[0])! >= major && Int(v[1])! >= minor
-        }
-
-        public func checkVersionCompatible(major: Int, minor: Int=0, micro: Int) -> Bool{
-            let v = self.redis_version.components(separatedBy: ".")
-            return Int(v[0])! >= major && Int(v[1])! >= minor && Int(v[2])! >= micro
-        }
-    }
-    
-}
